Fixed all the align and drive to commands
Drive to cube node needs to be tuned, the others should work okay, our cone nodes are all different distances which is an issue.

Removed the old generic align and drive commands in favor of the dedicated ones.

We should probably hard code the alignment off-sets somehow instead of passing them.